Add config param for autocomplete filtering
