+++ cmd/tempo/app/app.go
@@ -51,8 +51,9 @@ var (
 		nil,
 	)
 
-	statFeatureEnabledAuth         = usagestats.NewInt("feature_enabled_auth_stats")
-	statFeatureEnabledMultitenancy = usagestats.NewInt("feature_enabled_multitenancy")
+	statFeatureEnabledAuth                  = usagestats.NewInt("feature_enabled_auth_stats")
+	statFeatureEnabledMultitenancy          = usagestats.NewInt("feature_enabled_multitenancy")
+	statFeatureAutocompleteFilteringEnabled = usagestats.NewInt("feature_enabled_autocomplete_filtering")
 )
 
 // App is the root datastructure.
@@ -98,6 +99,11 @@ func New(cfg Config) (*App, error) {
 	statFeatureEnabledMultitenancy.Set(0)
 	if cfg.MultitenancyEnabled {
 		statFeatureEnabledMultitenancy.Set(1)
+	}
+
+	statFeatureAutocompleteFilteringEnabled.Set(0)
+	if cfg.AutocompleteFilteringEnabled {
+		statFeatureAutocompleteFilteringEnabled.Set(1)
 	}
 
 	app.setupAuthMiddleware()
