+++ cmd/tempo/app/config.go
@@ -27,12 +27,13 @@ import (
 
 // Config is the root config for App.
 type Config struct {
-	Target                 string `yaml:"target,omitempty"`
-	AuthEnabled            bool   `yaml:"auth_enabled,omitempty"`
-	MultitenancyEnabled    bool   `yaml:"multitenancy_enabled,omitempty"`
-	HTTPAPIPrefix          string `yaml:"http_api_prefix"`
-	UseOTelTracer          bool   `yaml:"use_otel_tracer,omitempty"`
-	EnableGoRuntimeMetrics bool   `yaml:"enable_go_runtime_metrics,omitempty"`
+	Target                       string `yaml:"target,omitempty"`
+	AuthEnabled                  bool   `yaml:"auth_enabled,omitempty"`
+	MultitenancyEnabled          bool   `yaml:"multitenancy_enabled,omitempty"`
+	HTTPAPIPrefix                string `yaml:"http_api_prefix"`
+	UseOTelTracer                bool   `yaml:"use_otel_tracer,omitempty"`
+	EnableGoRuntimeMetrics       bool   `yaml:"enable_go_runtime_metrics,omitempty"`
+	AutocompleteFilteringEnabled bool   `yaml:"autocomplete_filtering_enabled"`
 
 	Server          server.Config           `yaml:"server,omitempty"`
 	InternalServer  internalserver.Config   `yaml:"internal_server,omitempty"`
@@ -67,6 +68,7 @@ func (c *Config) RegisterFlagsAndApplyDefaults(prefix string, f *flag.FlagSet) {
 	f.StringVar(&c.HTTPAPIPrefix, "http-api-prefix", "", "String prefix for all http api endpoints.")
 	f.BoolVar(&c.UseOTelTracer, "use-otel-tracer", false, "Set to true to replace the OpenTracing tracer with the OpenTelemetry tracer")
 	f.BoolVar(&c.EnableGoRuntimeMetrics, "enable-go-runtime-metrics", false, "Set to true to enable all Go runtime metrics")
+	f.BoolVar(&c.AutocompleteFilteringEnabled, "autocomplete-filtering.enabled", false, "Set to true to enable autocomplete filtering")
 
 	// Server settings
 	flagext.DefaultValues(&c.Server)
