+++ integration/e2e/api_test.go
@@ -22,7 +22,7 @@ func TestSearchTagValuesV2(t *testing.T) {
 	defer s.Close()
 
 	require.NoError(t, util.CopyFileToSharedDir(s, configAllInOneLocal, "config.yaml"))
-	tempo := util.NewTempoAllInOne()
+	tempo := util.NewTempoAllInOne("-autocomplete-filtering.enabled=true")
 	require.NoError(t, s.StartAndWaitReady(tempo))
 
 	jaegerClient, err := util.NewJaegerGRPCClient(tempo.Endpoint(14250))
