+++ integration/util.go
@@ -9,7 +9,6 @@ import (
 	"os"
 	"path/filepath"
 	"strconv"
-	"strings"
 	"testing"
 	"time"
 
@@ -30,19 +29,7 @@ const (
 	image = "tempo:latest"
 )
 
-// GetExtraArgs returns the extra args to pass to the Docker command used to run Tempo.
-func GetExtraArgs() []string {
-	// Get extra args from the TEMPO_EXTRA_ARGS env variable
-	// falling back to an empty list
-	if os.Getenv("TEMPO_EXTRA_ARGS") != "" {
-		return strings.Fields(os.Getenv("TEMPO_EXTRA_ARGS"))
-	}
-
-	return nil
-}
-
-func buildArgsWithExtra(args []string) []string {
-	extraArgs := GetExtraArgs()
+func buildArgsWithExtra(args, extraArgs []string) []string {
 	if len(extraArgs) > 0 {
 		return append(extraArgs, args...)
 	}
@@ -50,9 +37,9 @@ func buildArgsWithExtra(args []string) []string {
 	return args
 }
 
-func NewTempoAllInOne() *e2e.HTTPService {
+func NewTempoAllInOne(extraArgs ...string) *e2e.HTTPService {
 	args := []string{"-config.file=" + filepath.Join(e2e.ContainerSharedDir, "config.yaml")}
-	args = buildArgsWithExtra(args)
+	args = buildArgsWithExtra(args, extraArgs)
 
 	s := e2e.NewHTTPService(
 		"tempo",
