+++ pkg/traceql/engine_test.go
@@ -447,26 +447,22 @@ func TestExecuteTagValues(t *testing.T) {
 	}
 
 	type testCase struct {
-		name           string
-		req            *tempopb.SearchTagValuesRequest
-		expectedValues []tempopb.TagValue
+		name             string
+		attribute, query string
+		expectedValues   []tempopb.TagValue
 	}
 
 	for _, tc := range []testCase{
 		{
-			name: "scoped param, no query",
-			req: &tempopb.SearchTagValuesRequest{
-				TagName: "resource.service.name",
-				Query:   "{}",
-			},
+			name:           "scoped param, no query",
+			attribute:      "resource.service.name",
+			query:          "{}",
 			expectedValues: []tempopb.TagValue{{Type: "String", Value: "my-service"}},
 		},
 		{
-			name: "intrinsic param, no query",
-			req: &tempopb.SearchTagValuesRequest{
-				TagName: "name",
-				Query:   "{}",
-			},
+			name:      "intrinsic param, no query",
+			attribute: "name",
+			query:     "{}",
 			expectedValues: []tempopb.TagValue{
 				{Type: "String", Value: "HTTP GET /status"},
 				{Type: "String", Value: "HTTP POST /api/v1/users"},
@@ -474,19 +470,15 @@ func TestExecuteTagValues(t *testing.T) {
 			},
 		},
 		{
-			name: "scoped param, with query",
-			req: &tempopb.SearchTagValuesRequest{
-				TagName: "span.http.method",
-				Query:   `{ span.http.target = "/api/v1/users" }`,
-			},
+			name:           "scoped param, with query",
+			attribute:      "span.http.method",
+			query:          `{ span.http.target = "/api/v1/users" }`,
 			expectedValues: []tempopb.TagValue{{Type: "String", Value: "POST"}},
 		},
 		{
-			name: "intrinsic param, with query",
-			req: &tempopb.SearchTagValuesRequest{
-				TagName: "name",
-				Query:   `{ span.http.target = "/api/v1/users" }`,
-			},
+			name:           "intrinsic param, with query",
+			attribute:      "name",
+			query:          `{ span.http.target = "/api/v1/users" }`,
 			expectedValues: []tempopb.TagValue{{Type: "String", Value: "HTTP POST /api/v1/users"}},
 		},
 	} {
@@ -495,7 +487,9 @@ func TestExecuteTagValues(t *testing.T) {
 			distinctValues := util.NewDistinctValueCollector[tempopb.TagValue](100_000, func(v tempopb.TagValue) int { return len(v.Type) + len(v.Value) })
 			cb := func(v Static) bool { return distinctValues.Collect(tempopb.TagValue{Type: "String", Value: v.S}) }
 
-			assert.NoError(t, e.ExecuteTagValues(context.Background(), tc.req, cb, mockSpansetFetcher()))
+			tag, err := ParseIdentifier(tc.attribute)
+			assert.NoError(t, err)
+			assert.NoError(t, e.ExecuteTagValues(context.Background(), tag, tc.query, cb, mockSpansetFetcher()))
 			values := distinctValues.Values()
 			sort.Slice(values, func(i, j int) bool {
 				return values[i].Value < values[j].Value
@@ -504,91 +498,4 @@ func TestExecuteTagValues(t *testing.T) {
 		})
 	}
 
-}
-
-func TestExtractMatchers(t *testing.T) {
-	testCases := []struct {
-		name, query, expected string
-	}{
-		{
-			name:     "empty query",
-			query:    "",
-			expected: "{}",
-		},
-		{
-			name:     "empty query with spaces",
-			query:    " { } ",
-			expected: "{}",
-		},
-		{
-			name:     "simple query",
-			query:    `{.service_name = "foo"}`,
-			expected: `{.service_name = "foo"}`,
-		},
-		{
-			name:     "incomplete query",
-			query:    `{ .http.status_code = 200 && .http.method = }`,
-			expected: "{.http.status_code = 200}",
-		},
-		{
-			name:     "invalid query",
-			query:    "{ 2 = .b ",
-			expected: "{}",
-		},
-		{
-			name:     "long query",
-			query:    `{.service_name = "foo" && .http.status_code = 200 && .http.method = "GET" && .cluster = }`,
-			expected: `{.service_name = "foo" && .http.status_code = 200 && .http.method = "GET"}`,
-		},
-		{
-			name:     "query with duration a boolean",
-			query:    `{ duration > 5s && .success = true && .cluster = }`,
-			expected: `{duration > 5s && .success = true}`,
-		},
-		{
-			name:     "query with three selectors with AND",
-			query:    `{ .foo = "bar" && .baz = "qux" } && { duration > 1s } || { .foo = "bar" && .baz = "qux" }`,
-			expected: "{}",
-		},
-		{
-			name:     "query with OR conditions",
-			query:    `{ (.foo = "bar" || .baz = "qux") && duration > 1s }`,
-			expected: "{}",
-		},
-		{
-			name:     "query with multiple selectors and pipelines",
-			query:    `{ .foo = "bar" && .baz = "qux" } && { duration > 1s } || { .foo = "bar" && .baz = "qux" } | count() > 4`,
-			expected: "{}",
-		},
-		{
-			name:     "query with slash in value",
-			query:    `{ span.http.target = "/api/v1/users" }`,
-			expected: `{span.http.target = "/api/v1/users"}`,
-		},
-	}
-	for _, tc := range testCases {
-		t.Run(tc.name, func(t *testing.T) {
-			assert.Equal(t, tc.expected, extractMatchers(tc.query))
-		})
-	}
-}
-
-func BenchmarkExtractMatchers(b *testing.B) {
-	queries := []string{
-		`{.service_name = "foo"}`,
-		`{.service_name = "foo" && .http.status_code = 200}`,
-		`{.service_name = "foo" && .http.status_code = 200 && .http.method = "GET"}`,
-		`{.service_name = "foo" && .http.status_code = 200 && .http.method = "GET" && .http.url = "/foo"}`,
-		`{.service_name = "foo" && .cluster = }`,
-		`{.service_name = "foo" && .http.status_code = 200 && .cluster = }`,
-		`{.service_name = "foo" && .http.status_code = 200 && .http.method = "GET" && .cluster = }`,
-		`{.service_name = "foo" && .http.status_code = 200 && .http.method = "GET" && .http.url = "/foo" && .cluster = }`,
-	}
-	for _, query := range queries {
-		b.Run(query, func(b *testing.B) {
-			for i := 0; i < b.N; i++ {
-				_ = extractMatchers(query)
-			}
-		})
-	}
 }
