+++ tempodb/encoding/v2/backend_block.go
@@ -153,6 +153,10 @@ func (b *BackendBlock) SearchTagValues(ctx context.Context, tag string, cb commo
 	return common.ErrUnsupported
 }
 
+func (b *BackendBlock) SearchTagValuesV2(context.Context, traceql.Attribute, common.TagCallbackV2, common.SearchOptions) error {
+	return common.ErrUnsupported
+}
+
 func (b *BackendBlock) Fetch(context.Context, traceql.FetchSpansRequest, common.SearchOptions) (traceql.FetchSpansResponse, error) {
 	return traceql.FetchSpansResponse{}, common.ErrUnsupported
 }
