+++ tempodb/encoding/vparquet/block_search_tags_test.go
@@ -139,7 +139,7 @@ func TestBackendBlockSearchTagValuesV2(t *testing.T) {
 			return false
 		}
 
-		err := block.searchTagValuesV2(ctx, tc.tag, cb, common.DefaultSearchOptions())
+		err := block.SearchTagValuesV2(ctx, tc.tag, cb, common.DefaultSearchOptions())
 		require.NoError(t, err, tc.tag)
 		require.Equal(t, tc.vals, got, "tag=%v", tc.tag)
 	}
